requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_cchykh5sgr/croots/recipe/asgiref_1661287542702/work
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
Django @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_eeor5o9jj8/croots/recipe/django_1661292534510/work
django-filter==23.3
djangorestframework==3.14.0
fair-esm==2.0.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
idna==3.4
Jinja2==3.1.2
lightning-utilities==0.9.0
Markdown==3.5.1
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
python-dateutil==2.8.2
pytorch-lightning==2.1.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse @ file:///C:/b/abs_88361ub_qu/croot/sqlparse_1690904577514/work
sympy==1.12
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3==2.0.7
yarl==1.9.2
